requirements.txt
alembic==1.15.2
aniso8601==10.0.0
argh==0.31.3
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pilkit==3.0
pillow==11.2.1
pillowcase==1.1.8
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
